fix: integrate features(login & main ui)
