+++ fortuna_android/app/src/main/java/com/example/fortuna_android/ui/HomeFragment.kt
@@ -3,6 +3,7 @@ package com.example.fortuna_android.ui
 import android.Manifest
 import android.content.pm.PackageManager
 import android.os.Bundle
+import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -20,6 +21,7 @@ class HomeFragment : Fragment() {
     private val binding get() = _binding!!
 
     private lateinit var fortuneViewModel: FortuneViewModel
+    private var lastFortuneResult: String? = null
 
     companion object {
         private const val TAG = "HomeFragment"
@@ -29,15 +31,21 @@ class HomeFragment : Fragment() {
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
+        Log.d(TAG, "onCreateView: HomeFragment view created")
         _binding = FragmentHomeBinding.inflate(inflater, container, false)
         return binding.root
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
+        Log.d(TAG, "onViewCreated: Setting up HomeFragment")
 
         // Initialize ViewModel (activity-scoped to persist across fragments)
         fortuneViewModel = ViewModelProvider(requireActivity())[FortuneViewModel::class.java]
+        Log.d(TAG, "ViewModel initialized")
+
+        // Initialize lastFortuneResult with current ViewModel state to prevent duplicate toasts
+        lastFortuneResult = fortuneViewModel.fortuneResult.value
 
         // Set up button click listeners
         binding.btnHome.setOnClickListener {
@@ -55,13 +63,13 @@ class HomeFragment : Fragment() {
     private fun setupObservers() {
         // Observe fortune result
         fortuneViewModel.fortuneResult.observe(viewLifecycleOwner) { fortune ->
+            Log.d(TAG, "Fortune result received: ${if (fortune != null) "Success" else "Null"}")
             updateFortuneText(fortune)
 
-            // Show success toast when fortune is generated
-            if (fortune != null && fortuneViewModel.isLoading.value == false) {
-                if (isAdded) {
-                    Toast.makeText(requireContext(), "Fortune generated successfully!", Toast.LENGTH_SHORT).show()
-                }
+            // Show success toast only for new fortune results, not when returning to fragment
+            if (fortune != null && fortune != lastFortuneResult && isAdded) {
+                Toast.makeText(requireContext(), "Fortune generated successfully!", Toast.LENGTH_SHORT).show()
+                lastFortuneResult = fortune
             }
         }
 
@@ -107,6 +115,7 @@ class HomeFragment : Fragment() {
         } else {
             binding.btnFortune.isEnabled = true
             binding.btnFortune.text = "Fortune Button"
+            Log.d(TAG, "Loading UI state cleared - button enabled")
             // Don't change the text when loading stops - let fortune result handle it
         }
     }
@@ -122,7 +131,9 @@ class HomeFragment : Fragment() {
 
     private fun checkPermissions() {
         // Check if fragment is still attached
-        if (!isAdded) return
+        if (!isAdded) {
+            return
+        }
 
         val requiredPermissions = arrayOf(
             Manifest.permission.CAMERA,
@@ -135,17 +146,19 @@ class HomeFragment : Fragment() {
 
         if (missingPermissions.isEmpty()) {
             // All permissions granted - navigate to camera
+            Log.d(TAG, "All permissions granted - navigating to camera")
             if (isAdded) {
                 Toast.makeText(requireContext(), "All permissions granted! Ready to proceed.", Toast.LENGTH_SHORT).show()
                 findNavController().navigate(R.id.cameraFragment)
             }
         } else {
             // Some permissions missing - request them via MainActivity
+            Log.d(TAG, "Missing permissions: ${missingPermissions.joinToString()}")
             if (isAdded) {
                 Toast.makeText(requireContext(), "Some permissions missing. Requesting permissions...", Toast.LENGTH_SHORT).show()
             }
             if (activity is MainActivity) {
-                (activity as MainActivity).requestPermissions()
+                (activity as? MainActivity)?.requestPermissions()
             }
         }
     }
